feat: Data Classification 및 Test 완료 (User, Package, Schedule, Banner)

+++ lib/data/repository/user_repository_impl.dart
@@ -42,7 +42,7 @@ class UserRepositoryImpl implements UserRepository {
                                 title: e.title,
                                 content: e.content,
                                 imageUrl: e.imageUrl,
-                                order: e.order as int,
+                                order: e.order,
                               ))
                           .toList(),
                       createdAt: package.createdAt,
